Bugs with the fewer hours
Made bandaid-like bug fixes but it doesn’t really fix the whole
problem, which is that the fewer than 7 hours of sleep condition also
works for the GREATER than 7 hours of sleep condition teehee. We’ll
discuss tomorrow but we’re going to bed (as an angry kitten has
suggested we do).